修改了处理顺序逻辑，理论上加快了速度

+++ dataprocessing/read_el_from_sp3.py
@@ -5,12 +5,14 @@ import datetime
 from pyproj import Transformer
 import utils
 
-def read_el_from_sp3(station_name, inputdir, output_folder, satellite_code, station_llh, start_date_str, end_date_str):
-    utils.create_directory_if_not_exists(output_folder)
+def read_el_from_sp3(station_name, inputdir, output_folder, satellite_range, station_llh, start_date_str, end_date_str):
+    utils.create_directory_if_not_exists(output_folder) # 如果没有文件夹，创建文件夹
 
     # 解析 start_date 和 end_date
-    start_date = pd.to_datetime(start_date_str, format='%Y%j')
-    end_date = pd.to_datetime(end_date_str, format='%Y%j')
+    start_date = pd.to_datetime(start_date_str, format='%Y%j')  # 处理数据开始日期
+    end_date = pd.to_datetime(end_date_str, format='%Y%j')  # 处理数据结束日期
+
+    # 加入对inputdir文件排序
 
     # 遍历data文件夹下的所有SP3文件，并进行插值
     for file in os.listdir(inputdir):
@@ -29,19 +31,11 @@ def read_el_from_sp3(station_name, inputdir, output_folder, satellite_code, stat
                 continue
 
             filename = os.path.join(inputdir, file)
-            outfilename = f'elevation_{station_name}_{file[11:18]}_{satellite_code}.csv'
-            full_path = os.path.join(output_folder, outfilename)
 
-            # 检查输出文件是否已经存在
-            if os.path.exists(full_path):
-                print(f"File {full_path} already exists. Skipping...")
-                continue
-
-            result = calculate_satellite_elevations(filename, station_llh, output_folder, satellite_code, full_path)
-            print(f"Execution result for {outfilename}:", "Success" if result == 0 else "Failure")
+            result = calculate_satellite_elevations(filename, station_llh, output_folder, satellite_range, station_name)
 
 
-def calculate_satellite_elevations(filename, station_llh, output_folder,satellite_code,full_path):
+def calculate_satellite_elevations(filename, station_llh, output_folder,satellite_range, station_name):
     """
     计算卫星仰角并将结果保存到CSV文件。
 
@@ -55,50 +49,64 @@ def calculate_satellite_elevations(filename, station_llh, output_folder,satellit
     try:
         # 第一步：解析sp3文件，获取时间戳-卫星位置对应的15min间隔数据
         [epochs, sat_data] = parse_sp3_file(filename)
-        # 只保留指定的卫星
-        if satellite_code not in sat_data:
-            raise ValueError(f"Satellite {satellite_code} not found in the SP3 file.")
 
-        sat_data = {satellite_code: sat_data[satellite_code]}
+        # 遍历satellite_range中的每一个卫星，执行下面的内容
+        for satellite_code in satellite_range:
+            # 只保留指定的卫星
+            if satellite_code not in sat_data:
+                raise ValueError(f"Satellite {satellite_code} not found in the SP3 file.")
 
-        # 第二步，对epochs和sat_data进行插值
-        new_epochs = create_new_epochs(epochs)
+            sat_single_data = {satellite_code: sat_data[satellite_code]}
 
-        # 对sat_data每个卫星的值进行插值
-        for sat_id in sat_data:
-            for coord in ['x', 'y', 'z']:
-                df = pd.DataFrame({'epoch': epochs, 'values': sat_data[sat_id][coord]})
-                df.set_index('epoch', inplace=True)
-                new_df = df.resample('30S').interpolate(method='linear')
-                new_values = new_df.reindex(new_epochs).values.flatten()
-                sat_data[sat_id][coord] = new_values
+            # 第二步，对epochs和sat_data进行插值
+            new_epochs = create_new_epochs(epochs)
 
-        # 第三步：对每个时间戳，计算卫星的仰角
-        # 计算station的ecef坐标
-        sta_x, sta_y, sta_z = llh_to_ecef(station_llh[0], station_llh[1], station_llh[2])
+            # 对sat_data每个卫星的值进行插值
+            for sat_id in sat_single_data:
+                for coord in ['x', 'y', 'z']:
+                    df = pd.DataFrame({'epoch': epochs, 'values': sat_single_data[sat_id][coord]})
+                    df.set_index('epoch', inplace=True)
+                    new_df = df.resample('30s').interpolate(method='linear')
+                    new_values = new_df.reindex(new_epochs).values.flatten()
+                    sat_single_data[sat_id][coord] = new_values
 
-        # 遍历每颗卫星
-        for sat_id in sat_data:
-            elevation_data = []
+            # 第三步：对每个时间戳，计算卫星的仰角
+            # 计算station的ecef坐标
+            sta_x, sta_y, sta_z = llh_to_ecef(station_llh[0], station_llh[1], station_llh[2])
 
-            # 遍历每个时间戳，求解那一时刻的仰角
-            for i, epoch in enumerate(new_epochs):
-                sat_x, sat_y, sat_z = sat_data[sat_id]['x'][i], sat_data[sat_id]['y'][i], sat_data[sat_id]['z'][i]
-                sat_ecef = [sat_x, sat_y, sat_z]
+            # 遍历每颗卫星
+            for sat_id in sat_single_data:
+                elevation_data = []
 
-                # 计算仰角
-                elevation = calc_el_angle(sat_ecef, [sta_x, sta_y, sta_z], station_llh[0], station_llh[1])
+                # 遍历每个时间戳，求解那一时刻的仰角
+                for i, epoch in enumerate(new_epochs):
+                    sat_x, sat_y, sat_z = sat_single_data[sat_id]['x'][i], sat_single_data[sat_id]['y'][i], sat_single_data[sat_id]['z'][i]
+                    sat_ecef = [sat_x, sat_y, sat_z]
 
-                # 将当前的epoch和仰角对应存储到变量中
-                elevation_data.append({'Timestamp': epoch, 'elevation': elevation})
+                    # 计算仰角
+                    elevation = calc_el_angle(sat_ecef, [sta_x, sta_y, sta_z], station_llh[0], station_llh[1])
 
-            # 将当前卫星的仰角结果保存到一个CSV文件中
-            df = pd.DataFrame(elevation_data)
-            df.to_csv(full_path, index=False)
+                    # 将当前的epoch和仰角对应存储到变量中
+                    elevation_data.append({'Timestamp': epoch, 'elevation': elevation})
 
-            del elevation_data
+                # 将当前卫星的仰角结果保存到一个CSV文件中
+                df = pd.DataFrame(elevation_data)
 
-        return 0  # 执行成功
+                filename_nopath = os.path.basename(filename)
+                outfilename = f'elevation_{station_name}_{filename_nopath[11:18]}_{satellite_code}.csv'
+                full_path = os.path.join(output_folder, outfilename)
+
+                # 检测输出文件是否已经存在
+                if os.path.exists(full_path):
+                    print(f"File {full_path} already exists. Skipping...")
+                    continue
+                else:
+                    df.to_csv(full_path, index=False)
+                    print(f"Execution result for {full_path}:", "Success")
+
+                del elevation_data
+
+            return 0  # 执行成功
     except Exception as e:
         print(f"An error occurred: {e}")
         return 1  # 执行失败
@@ -159,7 +167,7 @@ def create_new_epochs(epochs):
     # Create new epochs
     start_time = epochs[0]
     end_time = epochs[-1] - pd.Timedelta(minutes=0, seconds=30)
-    new_epochs = pd.date_range(start=start_time, end=end_time, freq='30S')
+    new_epochs = pd.date_range(start=start_time, end=end_time, freq='30s')
 
     return new_epochs
